Display comments count in card
