+++ src/api/comments-api-service.js
@@ -31,7 +31,6 @@ export default class CommentsApiService extends ApiService {
       url: `comments/${commentId}`,
       method: Method.DELETE,
     });
-
     return response;
   };
